+++ st2common/st2common/config.py
@@ -68,7 +68,9 @@ def register_opts(ignore_errors=False):
 
     system_opts = [
         cfg.StrOpt('base_path', default='/opt/stackstorm',
-                   help='Base path to all st2 artifacts.')
+                   help='Base path to all st2 artifacts.'),
+        cfg.ListOpt('admin_users', default=[],
+                    help='A list of usernames for users which should have admin privileges')
     ]
     do_register_opts(system_opts, 'system', ignore_errors)
 
@@ -112,7 +114,9 @@ def register_opts(ignore_errors=False):
         cfg.ListOpt('excludes', default='',
                     help='Exclusion list of loggers to omit.'),
         cfg.BoolOpt('redirect_stderr', default=False,
-                    help='Controls if stderr should be redirected to the logs.')
+                    help='Controls if stderr should be redirected to the logs.'),
+        cfg.BoolOpt('mask_secrets', default=True,
+                    help='True to mask secrets in the log files.')
     ]
     do_register_opts(log_opts, 'log', ignore_errors)