+++ CHANGELOG.rst
@@ -23,6 +23,8 @@ in development
 * API server now gracefully shuts down on SIGINT (CTRL-C). (improvement)
 * Fix a bug with with reinstalling a pack with no existing config - only try to move the config
   file over if it exists. (bug fix)
+* Support for masking secret parameters in the API responses. Secret parameters can only be viewed
+  through the API by admin users. (new-feature)
 
 0.11.3 - June 16, 2015
 ----------------------